Update patch set 1
Patch Set 1:

While I understand all of your points, and I could care less since I do not use Omnis standard kernel and almost nobody does, one thing remains: we are on QCOM and QCOM designed it the way it is, it's intended and not an error.

The proper way would be a system app that iterates over the cpus. This requires working uevents that reapply file perms on hotplug which are non-functional at this time.

Patch-set: 1