[DO NOT MERGE] Squashed commit of Marc van der Wielen's patches
I don't like "test this other branch" as a test methodology,
so I'm going to put up a squashed merge for review.  If it goes
well, we'll push the individual commits.  Same as the other
mega-merge reviews.  :)

(I removed the individual commits from the list, the commit message was, after looking at it, WAY too long...)

For detail, look at history with HEAD at
https://github.com/myfluxi/kernel_msm/commit/e314de146be9ce653cc2d0265292479b304bd3c8

Change-Id: I229762fccc15e7861f7890f4cf2f0c9a74275ae3